+++ src/main.ts
@@ -1,6 +1,8 @@
 import { createServer } from "http";
 import { createYoga } from "graphql-yoga";
-import { schema } from "./schema";
+import { makeExecutableSchema } from "@graphql-tools/schema";
+import { resolvers } from "./resolvers";
+import { typeDefinitions } from "./schema/type-definitions";
 
 function main() {
   const yoga = createYoga({ schema });
@@ -9,5 +11,10 @@ function main() {
     console.info("Server is running on http://localhost:4001/graphql");
   });
 }
+
+export const schema = makeExecutableSchema({
+  resolvers: [resolvers],
+  typeDefs: [typeDefinitions],
+});
 
 main();
