+++ src/mock-data/mock-data.ts
@@ -1,0 +1,9 @@
+import { Link } from "../types";
+
+export const links: Link[] = [
+  {
+    id: "link-0",
+    url: "https://graphql-yoga.com",
+    description: "The easiest way of setting up a GraphQL server",
+  },
+];
