+++ src/resolvers/queries/feed.ts
@@ -1,0 +1,3 @@
+import { links } from "../../mock-data/mock-data";
+
+export const feedResolver = () => links;
