+++ src/resolvers/queries/info.ts
@@ -1,0 +1,1 @@
+export const infoResolver = () => `This is the API of a Hackernews Clone`;
