+++ src/schema/type-definitions.ts
@@ -1,0 +1,16 @@
+export const typeDefinitions = /* GraphQL */ `
+  type Query {
+    info: String!
+    feed: [Link!]!
+  }
+
+  type Mutation {
+    postLink(url: String!, description: String!): Link!
+  }
+
+  type Link {
+    id: ID!
+    description: String!
+    url: String!
+  }
+`;
