+++ src/types/index.ts
@@ -1,0 +1,5 @@
+export type Link = {
+  id: string;
+  url: string;
+  description: string;
+};
